requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
jinja2==3.1.4
python-multipart==0.0.9
tensorflow-cpu==2.16.1
numpy==1.26.4
pillow==10.3.0
opencv-python-headless==4.9.0.80
motor==3.4.0
passlib[bcrypt]==1.7.4
